Fixed decoding for unknown instructions (#372)
* Added decoding of new token program instructions.
Fixed decoding for unknown instructions - should prevent exceptions if programs are updated and deployed with new instructions before solnet.

* Unit test.

* Version bump.

* Missing resource.

+++ src/Solnet.Programs/SystemProgramInstructions.cs
@@ -34,7 +34,7 @@ namespace Solnet.Programs
         /// <summary>
         /// Represents the instruction types for the <see cref="SystemProgram"/>.
         /// </summary>
-        internal enum Values : byte
+        internal enum Values : uint
         {
             /// <summary>
             /// Create a new account.
